feat: finish migrate
