+++ packages/module-user/src/client/RoleUsersManager.tsx
@@ -1,112 +1,41 @@
-import React, { useContext, useEffect, useMemo, useRef, useState } from 'react';
+import React, { useContext, useEffect, useMemo, useState } from 'react';
 import {
   BlockProvider,
   ExtendCollectionsProvider,
   FixedBlockWrapper,
-  getIdsWithChildren,
   RenderChildrenWithAssociationFilter,
-  ResourceActionProvider,
   SchemaComponent,
   SchemaComponentOptions,
   TableBlockContext,
-  useActionContext,
-  useAPIClient,
-  useCollectionRecordData,
-  useDataBlockRequest,
-  useDataBlockResource,
-  useRequest,
-  useResourceActionContext,
-  useTableBlockContext,
   useTableBlockParams,
+  withDynamicSchemaProps,
 } from '@tachybase/client';
 import { RolesManagerContext } from '@tachybase/module-acl/client';
 import { createForm, FormContext, useField, useFieldSchema } from '@tachybase/schema';
 
-import { App } from 'antd';
-
-import { useFilterActionProps } from './hooks';
+import {
+  useAddRoleUsers,
+  useBulkRemoveUsers,
+  useFilterActionProps,
+  useRemoveUser,
+  useRoleUsersProps,
+  useRoleUsersServiceProps,
+} from './hooks';
 import { useUsersTranslation } from './locale';
 import { getRoleUsersSchema, userCollection } from './schemas/users';
 
-const useRemoveUser = () => {
-  const api = useAPIClient();
-  const { role } = useContext(RolesManagerContext);
-  const record = useCollectionRecordData();
-  const { service } = useTableBlockContext();
-  return {
-    async onClick() {
-      await api.resource('roles.users', role?.name).remove({
-        values: [record['id']],
-      });
-      service?.refresh();
-    },
-  };
-};
-
-const useBulkRemoveUsers = () => {
-  const { t } = useUsersTranslation();
-  const { message } = App.useApp();
-  const api = useAPIClient();
-  const { service, field } = useTableBlockContext();
-  // const { state, setState, refresh } = useResourceActionContext();
-  const { role } = useContext(RolesManagerContext);
-
-  return {
-    async onClick() {
-      const selected = field?.data?.selectedRowKeys;
-      if (!selected?.length) {
-        message.warning(t('Please select users'));
-        return;
-      }
-      await api.resource('roles.users', role?.name).remove({
-        values: selected,
-      });
-      field.data.selectedRowKeys = [];
-      service?.refresh();
-    },
-  };
-};
-
-const useRoleUsersProps = (props) => {
-  const { role } = useContext(RolesManagerContext);
-
-  return {
-    ...props,
-    collection: userCollection,
-    action: 'listExcludeRole',
-    params: {
-      roleName: role?.name,
-    },
-  };
-};
-
-const RoleUsersTableBlockProvider = (props) => {
-  const { showIndex, dragSort, rowKey, fieldNames, collection, ...others } = props;
+export const RoleUsersTableBlockProvider = withDynamicSchemaProps((props) => {
+  const { showIndex, dragSort, rowKey, fieldNames, collection, service, ...others } = props;
   const field: any = useField();
   const { role } = useContext(RolesManagerContext);
   const params = useTableBlockParams(props);
-  const service = useRequest(
-    {
-      resource: 'roles.users',
-      resourceOf: role?.name,
-      action: 'list',
-    },
-    {
-      ready: !!role,
-    },
-  );
   useEffect(() => {
     service.run();
   }, [role]);
   const fieldSchema = useFieldSchema();
   const { treeTable } = fieldSchema?.['x-decorator-props'] || {};
   const [expandFlag, setExpandFlag] = useState(fieldNames ? true : false);
-  // const allIncludesChildren = useMemo(() => {
-  //   if (treeTable !== false) {
-  //     const keys = getIdsWithChildren(service?.data?.data);
-  //     return keys || [];
-  //   }
-  // }, [service?.loading]);
+
   let childrenColumnName = 'children';
   if (collection?.tree && treeTable !== false) {
     const f = collection.fields.find((f) => f.treeChildren);
@@ -143,58 +72,12 @@ const RoleUsersTableBlockProvider = (props) => {
       </FormContext.Provider>
     </SchemaComponentOptions>
   );
-};
+});
 
 export const RoleUsersManager: React.FC = () => {
   const { t } = useUsersTranslation();
   const { role } = useContext(RolesManagerContext);
-  // const service = useRequest(
-  //   {
-  //     resource: 'roles.users',
-  //     resourceOf: role?.name,
-  //     action: 'list',
-  //   },
-  //   {
-  //     ready: !!role,
-  //   },
-  // );
-  // useEffect(() => {
-  //   service.run();
-  // }, [role]);
-
-  // const selectedRoleUsers = useRef([]);
-  // const handleSelectRoleUsers = (_: number[], rows: any[]) => {
-  //   selectedRoleUsers.current = rows;
-  // };
-
-  const useAddRoleUsers = () => {
-    const { role } = useContext(RolesManagerContext);
-    const { message } = App.useApp();
-    const api = useAPIClient();
-    const { setVisible } = useActionContext();
-    const { field, service } = useTableBlockContext();
-    const { refresh } = useDataBlockRequest();
-    // const field = useField();
-    // const resource = useDataBlockResource;
-    return {
-      async onClick() {
-        const selected = field?.data?.selectedRowKeys;
-        if (!selected?.length) {
-          message.warning(t('Please select users'));
-          return;
-        }
-        await api.resource('roles.users', role?.name).add({
-          values: selected,
-        });
-        field.data.selectedRowKeys = [];
-        setVisible(false);
-        service.refresh();
-      },
-    };
-  };
-
   const schema = useMemo(() => getRoleUsersSchema(), [role]);
-  // const schema = getRoleUsersSchema();
 
   return (
     <ExtendCollectionsProvider collections={[userCollection]}>
@@ -207,6 +90,7 @@ export const RoleUsersManager: React.FC = () => {
           useAddRoleUsers,
           useRoleUsersProps,
           useFilterActionProps,
+          useRoleUsersServiceProps,
           t,
         }}
       />
